phantomnet-c2/requirements.txt
# Core web framework
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
Flask-CORS==4.0.0
flask-migrate
# Security and authentication
Werkzeug==3.0.3
cryptography==43.0.1
bcrypt==4.2.1
PyJWT==2.10.1

# HTTP requests and APIs
requests==2.32.3
urllib3==2.2.3

# Data processing
python-dateutil==2.9.0.post0

# Logging and debugging
colorlog==6.9.0

# Network scanning and enumeration
dnspython==2.6.1

# Note: python-nmap, virustotal-api, and abuseipdb are optional
# Install manually if needed:
# pip install python-nmap
# pip install virustotal-api
# pip install abuseipdb

# Image processing for screenshots
Pillow==11.0.0

# Development and testing
pytest==8.3.4
pytest-flask==1.3.0
pytest-mock==3.14.0
pytest-cov==6.0.0

# Code quality
black==24.10.0
flake8==7.1.1
mypy==1.13.0
isort==5.13.2

# Documentation
sphinx
sphinx-rtd-theme

# Production deployment
gunicorn==23.0.0
python-dotenv==1.0.1

# Additional security and utilities
psutil==6.1.0
pycryptodome==3.21.0
